ARM: tegra: Skip FDT and initrd relocation on 64-bit
U-Boot inspects the initrd_high and fdt_high environment variables and
skips relocation of the initial ramdisk and FDT, respectively, if these
variables are set to all-ones. The Tegra configuration sets these
variables to ffffffff in the default environment. However on 64-bit SoC
generations this will cause the all-ones match to fail, so extend these
variables to 64 ones on 64-bit SoCs.

Cc: Tom Warren <twarren@nvidia.com>
Signed-off-by: Thierry Reding <treding@nvidia.com>