ARM: tegra: p2371-2180: Add eth0 alias to DT
Note that the standard device tree bindings use "ethernet" as the alias
prefix. This slightly deviates from the standard bindings because U-Boot
looks for aliases with the "eth" prefix. This shouldn't matter, though,
since U-Boot doesn't typically pass its control DTB to the operating
system.

Signed-off-by: Thierry Reding <treding@nvidia.com>
---
An alternative would be to use the standard "ethernet" prefix in the
control DTB and instead change the Ethernet uclass' name, or perhaps
add an extra field to the uclass that defines the alias prefix if we
must keep the uclass' name as "eth".